requirements.txt
streamlit>=1.31.0
google-generativeai>=0.4.0